Add initializeProviders utility
